pyproject.toml
[tool.poetry]
name = "lms-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
flake8 = "^7.1.1"
mypy = "^1.13.0"
black = "^24.10.0"
isort = "^5.13.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
requests = "^2.32.3"
pillow = "^11.0.0"
ipython = "^8.29.0"
redis = "^5.2.0"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
